Pull out EllipsoidInsideTest in its own public class.
