+++ src/main/java/org/mastodon/detection/mamut/MamutDetectionCreatorFactories.java
@@ -12,9 +12,9 @@ import org.mastodon.revised.model.mamut.ModelGraph;
 import org.mastodon.revised.model.mamut.Spot;
 import org.mastodon.spatial.SpatialIndex;
 import org.mastodon.spatial.SpatioTemporalIndex;
+import org.mastodon.util.EllipsoidInsideTest;
 
 import net.imglib2.RealPoint;
-import net.imglib2.util.LinAlgHelpers;
 
 /**
  * Collection of {@link DetectionCreatorFactory}s suitable to be used with a
@@ -452,92 +452,6 @@ public class MamutDetectionCreatorFactories
 		public void postAddition()
 		{
 			graph.getLock().writeLock().unlock();
-		}
-	}
-
-	/**
-	 * Adapted from ScreenVertexMath.
-	 *
-	 */
-	private static class EllipsoidInsideTest
-	{
-
-		/** Holder for the first spot position. */
-		private final double[] pos1 = new double[ 3 ];
-
-		/** Holder for the second spot position. */
-		private final double[] pos2 = new double[ 3 ];
-
-		/** Holder for the covariance position. */
-		private final double[][] cov = new double[ 3 ][ 3 ];
-
-		/** Used to determines whether a spot contains a position. */
-		private final double[] diff = new double[ 3 ];
-
-		/** Used to determines whether a spot contains a position. */
-		private final double[] vn = new double[ 3 ];
-
-		/** Precision. */
-		private final double[][] P = new double[ 3 ][ 3 ];
-
-		/**
-		 * Returns <code>true</code> if the first spot contains the center of the second
-		 * one, or if the second spot contains the center of the first one.
-		 * 
-		 * @param s1
-		 *            the first spot.
-		 * @param s2
-		 *            the second spot.
-		 * @return <code>true</code> if a center of spot is included in the other spot.
-		 */
-		public boolean areCentersInside( final Spot s1, final Spot s2 )
-		{
-			s2.localize( pos2 );
-			if ( isPointInside( pos2, s1 ) )
-				return true;
-
-			s1.localize( pos2 );
-			return isPointInside( pos2, s2 );
-		}
-
-		/**
-		 * Returns <code>true</code> is the specified position lies inside the spot
-		 * ellipsoid.
-		 * 
-		 * @param pos
-		 *            the position to test.
-		 * @param spot
-		 *            the spot.
-		 * @return <code>true</code> if the position is inside the spot.
-		 */
-		public boolean isPointInside( final double[] pos, final Spot spot )
-		{
-			spot.localize( pos1 );
-			LinAlgHelpers.subtract( pos1, pos, diff );
-			spot.getCovariance( cov );
-			LinAlgHelpers.invertSymmetric3x3( cov, P );
-			LinAlgHelpers.mult( P, diff, vn );
-			final double d2 = LinAlgHelpers.dot( diff, vn );
-			return d2 < 1.;
-		}
-
-		/**
-		 * Returns <code>true</code> if the spot center is within the specified radius
-		 * around the specified position.
-		 * 
-		 * @param spot
-		 *            the spot to test.
-		 * @param pos
-		 *            the position.
-		 * @param radius
-		 *            the radius.
-		 * @return <code>true</code> if the spot center is within <code>radius</code> of
-		 *         <code>pos</code>.
-		 */
-		public boolean isCenterWithin( final Spot spot, final double[] pos, final double radius )
-		{
-			spot.localize( pos1 );
-			return LinAlgHelpers.squareDistance( pos1, pos ) < radius * radius;
 		}
 	}
